modified:   README.md 	modified:   src/main/java/com/hibernate_springboot_learning/hibernate_springboot_schema_sql/MainApplication.java 	renamed:    src/main/java/com/hibernate_springboot_learning/hibernate_springboot_schema_sql/BookstoreService.java -> src/main/java/com/hibernate_springboot_learning/hibernate_springboot_schema_sql/service/BookstoreService.java 	modified:   src/main/resources/application.properties 	new file:   src/main/resources/schema-mysql.sql
