+++ README.md
@@ -1,1 +1,4 @@
 Collection of best practices for Java persistence performance in Spring Boot applications
+
+
+docker run --name mysql-container -e MYSQL_ROOT_USERNAME=root -e MYSQL_ROOT_PASSWORD=root -p 3306:3306 -d mysql:8.0
