+++ src/main/resources/application.properties
@@ -4,9 +4,10 @@ spring.datasource.password=root
 
 spring.jpa.show-sql=true
 
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 
-spring.datasource.initialization-mode=always
-spring.datasource.platform=mysql
 
+spring.jpa.hibernate.ddl-auto=validate
+spring.sql.init.mode=always
+spring.sql.init.schema-locations=classpath:schema-mysql.sql
 spring.jpa.open-in-view=false
