+++ src/main/resources/schema-mysql.sql
@@ -1,0 +1,7 @@
+CREATE TABLE IF NOT EXISTS author (
+  id bigint(20) NOT NULL AUTO_INCREMENT,
+  age int(11) NOT NULL,
+  genre varchar(255) DEFAULT NULL,
+  name varchar(255) DEFAULT NULL,
+  PRIMARY KEY (id)
+);
